Fixed size_t in place of int to silence warning

+++ src/new_types/su3_op.hpp
@@ -128,7 +128,7 @@ namespace nissa
   CUDA_HOST_AND_DEVICE inline void color_summ_the_prod_complex(color a,const color b,const complex c) {for(size_t ic=0;ic<NCOL;ic++) complex_summ_the_prod(a[ic],b[ic],c);}
   CUDA_HOST_AND_DEVICE inline void color_subt_the_prod_complex(color a,const color b,const complex c) {for(size_t ic=0;ic<NCOL;ic++) complex_subt_the_prod(a[ic],b[ic],c);}
   
-  CUDA_HOST_AND_DEVICE inline void color_prod_color_delta(color a,const color b,const int c) {for(size_t ic=0;ic<NCOL;ic++) complex_prod_double(a[ic],b[ic],c==ic);}
+  CUDA_HOST_AND_DEVICE inline void color_prod_color_delta(color a,const color b,const size_t c) {for(size_t ic=0;ic<NCOL;ic++) complex_prod_double(a[ic],b[ic],c==ic);}
   
   inline void color_linear_comb(color a,const color b,const double cb,const color c,const double cc) {for(size_t ic=0;ic<NCOL;ic++) complex_linear_comb(a[ic],b[ic],cb,c[ic],cc);}
   
@@ -1410,7 +1410,7 @@ namespace nissa
   CUDA_HOST_AND_DEVICE inline void colorspinspin_prodassign_complex_conj(colorspinspin c,const complex f)
   {colorspinspin_prod_complex_conj(c,c,f);}
   
-  CUDA_HOST_AND_DEVICE inline void colorspinspin_prod_color_delta(colorspinspin out,const colorspinspin in,const int c)
+  CUDA_HOST_AND_DEVICE inline void colorspinspin_prod_color_delta(colorspinspin out,const colorspinspin in,const size_t c)
   {for(size_t ic=0;ic<NCOL;ic++) spinspin_prod_double(out[ic],in[ic],c==ic);}
   
   //colorspinspin summ
@@ -1449,7 +1449,7 @@ namespace nissa
   CUDA_HOST_AND_DEVICE inline void su3spinspin_prodassign_complex_conj(su3spinspin c,const complex f)
   {su3spinspin_prod_complex_conj(c,c,f);}
   
-  CUDA_HOST_AND_DEVICE inline void su3spinspin_prod_color_delta(su3spinspin out,const su3spinspin in,const int c)
+  CUDA_HOST_AND_DEVICE inline void su3spinspin_prod_color_delta(su3spinspin out,const su3spinspin in,const size_t c)
   {for(size_t ic=0;ic<NCOL;ic++) colorspinspin_prod_double(out[ic],in[ic],c==ic);}
   
   //su3spinspin summ
